get_bug_log: use BytesFeedParser for 'message' element
The Feedparser does not work reliably if the body contains non-ascii
characters. This can be noticed when trying to decode the body using
`msg.get_payload(decode=True).decode()` fails with UnicodeError.
Switching to the BytesFeedParser makes this work correctly.

For the BytesFeedParser the documentation strongly recommends always
specifying the policy. So let's do that, too.